+++ How_to_use_it.md
@@ -113,6 +113,42 @@ In questo modo è possibile conoscere come muovere le viti dello spider per rial
 	```
 NOTA: per maggiori informazioni fare riferimento anche alla seguente pagina wiki [WikiPage](https://redmine.ict.inaf.it/projects/adopt_oaa/wiki/MOTT-20200915)
 
+### Analisi dei requisiti ###
+Con i comandi
+```
+	
+	from m4.main import requirements_checker as rc
+    rc.analysis_req(data_file_path, zernike_vector_to_subtract, step=None, offset=None)
+```
+dove zernike_vector_to_subtract è il vettore degli zernike che si vuole sottrarre all'immagine robusta, step è la distanza tra le patches (necessaria 
+per il calcolo del raggio di curvatura e dell'rms interactuator) e piston distingue il metodo di creazione dell'immagine robusta: se è None l'immagine viene
+creata sottraendo tra loro i due cubi creati con la metà del numero di misure indicato, altrimenti al numero di misure indicato viene sottratta un'immagine di offset
+precedentemente salvata nella cartella con le misure.  
+NOTA: l'immagine di offset viene creata con i seguenti comandi
+```
+
+	from m4.analyzers import requirement_analyzer as ra
+	ra.imageOpticOffset(data_file_path, start, stop)
+```
+è possibile utilizzare una procedura standard che prevede la creazione di tre immagini robuste (create utilizzando 50, 100 e 300 file presenti nella cartella
+delle misure da analizzare). Le immagini robuste vengono analizzate e vengono automaticamente plottati e salvati i seguenti risultati:  
+	- slop
+	- differential piston
+	- radius of curvature 
+	- rms at the interactuator scale 31 mm
+	- rms at the interactuator scale 500 mm
+
+Nel caso in cui si voglia applicare l'analisi dei requisiti avando a disposizione una sola immagine utilizzare
+```
+
+	from m4.analyzers import requirement_analyzer as ra
+	slope = ra.test242(image, pscale)
+	diff_piston = ra.diffPiston(image)
+	roc = ra.test283(image, pscale, step)
+	rms31 = ra.test243(image, 0.015, pscale, step, n_patches)
+	rms500 = ra.est243(image, 0.1, pscale, step, n_patches)
+```
+
 ## Interferometro ##
 ### 4D PhaseCam 6110 ###
 I comandi da utilizzare sono:
